Prerequisite change for enabling `-Werror=sign-compare` across PyTorch repo

[ghstack-poisoned]

+++ caffe2/operators/generate_proposals_op_util_nms.h
@@ -50,7 +50,8 @@ std::vector<int> nms_cpu_upright(
   std::vector<int> keep;
   while (order.size() > 0) {
     // exit if already enough proposals
-    if (topN >= 0 && keep.size() >= static_cast<size_t>(topN)) {
+    // NOLINTNEXTLINE(clang-diagnostic-sign-compare)
+    if (topN >= 0 && keep.size() >= topN) {
       break;
     }
 
@@ -126,7 +127,7 @@ std::vector<int> soft_nms_cpu_upright(
   EArrXi pending = AsEArrXt(indices);
   while (pending.size() > 0) {
     // Exit if already enough proposals
-    if (topN >= 0 && keep.size() >= static_cast<unsigned>(topN)) {
+    if (topN >= 0 && keep.size() >= topN) {
       break;
     }
 
@@ -559,7 +560,8 @@ std::vector<int> nms_cpu_rotated(
   std::vector<int> keep;
   while (order.size() > 0) {
     // exit if already enough proposals
-    if (topN >= 0 && keep.size() >= static_cast<size_t>(topN)) {
+    // NOLINTNEXTLINE(clang-diagnostic-sign-compare)
+    if (topN >= 0 && keep.size() >= topN) {
       break;
     }
 
@@ -624,7 +626,7 @@ std::vector<int> soft_nms_cpu_rotated(
   EArrXi pending = AsEArrXt(indices);
   while (pending.size() > 0) {
     // Exit if already enough proposals
-    if (topN >= 0 && keep.size() >= static_cast<size_t>(topN)) {
+    if (topN >= 0 && keep.size() >= topN) {
       break;
     }
